Add reset punishment
