+++ src/main/java/io/github/connorhartley/guardian/internal/detections/JesusDetection.java
@@ -34,9 +34,11 @@ import io.github.connorhartley.guardian.detection.DetectionTypes;
 import io.github.connorhartley.guardian.detection.check.CheckType;
 import io.github.connorhartley.guardian.event.sequence.SequenceFinishEvent;
 import io.github.connorhartley.guardian.internal.checks.HorizontalSpeedCheck;
+import io.github.connorhartley.guardian.internal.contexts.player.PlayerLocationContext;
 import io.github.connorhartley.guardian.internal.punishments.CustomPunishment;
 import io.github.connorhartley.guardian.internal.punishments.KickPunishment;
 import io.github.connorhartley.guardian.internal.punishments.ReportPunishment;
+import io.github.connorhartley.guardian.internal.punishments.ResetPunishment;
 import io.github.connorhartley.guardian.internal.punishments.WarningPunishment;
 import io.github.connorhartley.guardian.punishment.Punishment;
 import io.github.connorhartley.guardian.storage.StorageConsumer;
@@ -101,6 +103,7 @@ public class JesusDetection extends Detection {
         this.configuration.create();
 
         this.plugin.getPunishmentController().bind(CustomPunishment.class, this);
+        this.plugin.getPunishmentController().bind(ResetPunishment.class, this);
         this.plugin.getPunishmentController().bind(WarningPunishment.class, this);
         this.plugin.getPunishmentController().bind(KickPunishment.class, this);
         this.plugin.getPunishmentController().bind(ReportPunishment.class, this);
@@ -297,6 +300,7 @@ public class JesusDetection extends Detection {
 
                 Map<String, Tuple<Double, Double>> punishmentLevels = new HashMap<>();
                 punishmentLevels.put("warn&default", Tuple.of(0.1, 1.0));
+                punishmentLevels.put("reset&default", Tuple.of(0.1, 1.0));
 
                 this.configPunishmentLevels = new StorageValue<>(new StorageKey<>("punishment-levels"),
                         new ConfigurationCommentDocument(45, " ")
