+++ src/main/java/io/github/connorhartley/guardian/internal/punishments/KickPunishment.java
@@ -54,8 +54,7 @@ public class KickPunishment implements PunishmentType {
 
     @Override
     public Optional<Detection> getDetection() {
-        if (this.detection == null) return Optional.empty();
-        return Optional.of(this.detection);
+        return Optional.ofNullable(this.detection);
     }
 
     @Override
